realtime/live-price-ingestor/requirements.txt
redis[hiredis]>=5.0
aiohttp>=3.9
python-dotenv>=1.0
